pyproject.toml
[project]
name = "risk-fda-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymilvus>=2.6.5",
    "python-dotenv>=1.2.1",
]
